move priority sampling tests to their own file
